LPD-25497 Moved all commerce thread locals under util packages
