+++ modules/apps/commerce/commerce-api/src/main/java/com/liferay/commerce/util/CommerceContextThreadLocal.java
@@ -3,8 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.commerce.context;
+package com.liferay.commerce.util;
 
+import com.liferay.commerce.context.CommerceContext;
 import com.liferay.petra.lang.CentralizedThreadLocal;
 import com.liferay.petra.lang.SafeCloseable;
