+++ modules/apps/commerce/commerce-api/src/main/java/com/liferay/commerce/util/CommerceReturnThreadLocal.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.commerce.order;
+package com.liferay.commerce.util;
 
 import com.liferay.petra.lang.CentralizedThreadLocal;
