+++ modules/apps/commerce/commerce-theme-minium/commerce-theme-minium-api/src/main/java/com/liferay/commerce/theme/minium/util/SiteInitializerDependencyResolverThreadLocal.java
@@ -3,8 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.commerce.theme.minium;
+package com.liferay.commerce.theme.minium.util;
 
+import com.liferay.commerce.theme.minium.SiteInitializerDependencyResolver;
 import com.liferay.petra.lang.CentralizedThreadLocal;
 
 /**
